refactor: move entity validator code into util
Refs #349
